clean up imports: factor out top-selectors/timelineWarnings
