+++ protocol-designer/src/file-data/selectors/commands.js
@@ -9,7 +9,7 @@ import { getAllWellsForLabware } from '../../constants'
 import * as StepGeneration from '../../step-generation'
 import { selectors as featureFlagSelectors } from '../../feature-flags'
 import { selectors as labwareIngredSelectors } from '../../labware-ingred/selectors'
-import * as stepFormSelectors from '../../step-forms/selectors'
+import { selectors as stepFormSelectors } from '../../step-forms'
 import type { BaseState, Selector } from '../../types'
 import type { StepIdType } from '../../form-types'
 import type {
